# Make Ahocorasick matching case insensitive