This PR fixes a bug that prevents secrets from being detected if the keyword or chunk data case varies. 

`strings.ToLower` doesn't seem to have a huge effect on performance either.

```
Benchmark git scan on the rails/rails repo

### NEW
real 27.18s
user 58.14s 
sys  5.47s 
memory:753328KB 
cpu 234%
      62

### OLD
real 27.45s
user 56.43s 
sys  5.50s 
memory:750160KB 
cpu 225%
      62
```

To demonstrate this bug fix, try finding a private key