@zricethezav 2023-07-25
@mcastorina I didn't see any, but I agree. Currently, some secrets like private keys are being detected because  `https://github.com/BobuSumisu/aho-corasick` does not have a case insensitive option like `https://github.com/petar-dambovaliev/aho-corasick` and therefore when `BEING PRIVATE KEY` is scanned it fails to match on `private key` (pkey detector keyword). 

Note that the performance hit in the test is a tradeoff for correctness